Fix password reset

+++ src/main/java/com/alpha/lainovo/dto/request/PublicationsDTO.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class PublicationsDTO {
     private String publicationsName;
-    private Float unitPrice;    
+    private Double unitPrice;
     private Integer stock;
     private String author;
     private String publisher;
